# sensor_hub/openapi.yaml
openapi: 3.1.0
info:
  title: Sensor Hub API
  version: 1.0.0
  description: >-
    API for Sensor Hub. Provides endpoints for managing sensors,
    querying raw readings, hourly averaged readings, updating application
    properties, managing users and RBAC, dashboards and more. Authentication is
    cookie-based with CSRF protection, as well as API Key for service-to-service
    communication. WebSocket endpoints are available for real-time updates of
    sensor readings and properties changes.
servers:
  - url: http://localhost:8080/api
security:
  - cookieAuth: []
    csrfToken: []
  - apiKeyAuth: []
paths:
...
  /properties:
    get:
      tags:
        - properties
      summary: Get current application properties
      description: >-
        Returns a flat object mapping property keys to their current values.
        Sensitive properties (such as OAuth secrets) are masked in
        responses as `"*****"` to avoid leaking secrets. Values are returned
        as strings.
      operationId: getProperties
      x-required-permission: view_properties
      responses:
        '200':
          description: Current properties
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertiesMap'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - properties
      summary: Update application properties
      description: >-
        Update one or more properties. Pass a JSON object where keys are the
        property names and values are the desired string values. To keep
        sensitive properties secret when you don't want to change them, pass
        the literal string `"*****"` as the value; the server will treat this
        as "do not change" for known sensitive keys.

        The update will be validated, applied to the in-memory configuration,
        and persisted to configuration files asynchronously. A broadcast is
        sent to the properties WebSocket topic after the update. Note: there is
        no separate endpoint for editing `smtp.properties` or `database.properties` —
        they are updated by key through this same API.
      operationId: updateProperties
      x-required-permission: manage_properties
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePropertiesRequest'
      responses:
        '202':
          description: Accepted — update applied asynchronously
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationAccepted'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /properties/ws:
    get:
      tags:
        - properties
      summary: WebSocket endpoint — subscribe to properties updates
      description: >-
        Upgrades the HTTP connection to a WebSocket and sends the current
        properties as a JSON object. Subsequent messages are sent when
        properties are updated (the server broadcasts the full properties map).
      operationId: propertiesWebSocket
      x-required-permission: view_properties
      responses:
        '101':
          description: Switching Protocols — connection upgraded to WebSocket
        '200':
          description: Non-upgrade response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-websocket:
        description: |
          WebSocket messages are JSON objects matching `PropertiesWebSocketMessage`.
          The first message after connect is the current properties map.
        messages:
          snapshot:
            schema:
              $ref: '#/components/schemas/PropertiesWebSocketMessage'
            example:
              timestamp: "2026-01-10T12:00:00Z"
              properties:
                smtp.user: "admin@example.com"
                smtp.recipient: "alerts@example.com"
                sensor.collection.interval: "300"

  # ============================================================================
  # Auth Endpoints
  # ============================================================================
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Human-readable error message.
      required:
        - message
...
    PropertiesMap:
      type: object
      description: |
        Flat map of configuration keys to values (all values as strings).
        Sensitive values are masked ("*****") in responses.
      additionalProperties:
        type: string
      example:
        smtp.user: "admin@example.com"
        smtp.recipient: "alerts@example.com"
        database.path: "data/sensor_hub.db"
        sensor.collection.interval: "300"
    UpdatePropertiesRequest:
      type: object
      description: Map of properties to update. Use the literal string "*****" to indicate no change for sensitive keys.
      additionalProperties:
        type: string
      example:
        smtp.user: "new-user@example.com"
        oauth.client_secret: "*****"
    PropertiesWebSocketMessage:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Server timestamp for the message.
        properties:
          $ref: '#/components/schemas/PropertiesMap'
      required:
        - timestamp
        - properties
    OperationAccepted:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # =========================================================================
    # Auth Schemas
    # =========================================================================
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sensor_hub_session
      description: Session cookie set after successful login. Configurable via AuthSessionCookieName.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF token returned from login, required for state-changing operations (POST, PUT, PATCH, DELETE).
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        API key for programmatic and CLI access.
        Exempt from CSRF validation. Create keys via the
        API Keys management endpoints.

  # Optional: group endpoints for UIs and tooling
...